Build plugin cleanup.
