+++ pb-coroutines-core/build.gradle.kts
@@ -1,5 +1,3 @@
 plugins {
-    kotlin("jvm")
-
-    `java-library`
+    id("core-config")
 }
