+++ pb-coroutines-kafka/build.gradle.kts
@@ -1,6 +1,5 @@
 plugins {
     kotlin("jvm")
-
     `java-library`
 }
